adding character list
