+++ src/components/character-list/character-list.module.scss
